Move `utils` module to `app`
