+++ app/routes/web/favourites.py
@@ -15,7 +15,6 @@ from app.common.database.repositories import (
     users
 )
 
-import utils
 import app
 
 router = APIRouter()
@@ -30,7 +29,7 @@ def add_favourite(
     if not (player := users.fetch_by_name(username, session)):
         raise HTTPException(401)
 
-    if not utils.check_password(password, player.bcrypt):
+    if not app.utils.check_password(password, player.bcrypt):
         raise HTTPException(401)
 
     if not status.exists(player.id):
@@ -69,7 +68,7 @@ def get_favourites(
     if not (player := users.fetch_by_name(username, session)):
         raise HTTPException(401)
 
-    if not utils.check_password(password, player.bcrypt):
+    if not app.utils.check_password(password, player.bcrypt):
         raise HTTPException(401)
 
     users.update(player.id, {'latest_activity': datetime.now()}, session)
