+++ app/routes/web/ratings.py
@@ -17,7 +17,6 @@ from app.common.database import (
 
 router = APIRouter()
 
-import utils
 import app
 
 @router.get('/osu-rate.php')
@@ -31,7 +30,7 @@ def rate(
     if not (player := users.fetch_by_name(username, session)):
         return 'auth fail'
 
-    if not utils.check_password(password, player.bcrypt):
+    if not app.utils.check_password(password, player.bcrypt):
         return 'auth fail'
 
     if not status.exists(player.id):
